organise the flows
