+++ intelli/test/integration/test_flow_sequence.py
@@ -4,7 +4,7 @@ from intelli.flow.agents.agent import Agent
 from intelli.flow.input.task_input import TextTaskInput
 from intelli.flow.processors.basic_processor import TextProcessor
 from intelli.flow.sequence_flow import SequenceFlow
-from intelli.flow.task import Task
+from intelli.flow.tasks.task import Task
 from dotenv import load_dotenv
 
 load_dotenv()
